143:     - commented unused AuthService     - move auth logic from header to appropriate component     - added possibility to open actor from welcome page
