+++ src/app/app.component.html
@@ -4,40 +4,6 @@
 
     <exp-header></exp-header>
 
-    <!-- AUTH -->
-    AUTH
-    <div style="display: flex; color: #000000; margin: 100px auto;">
-      <!-- this header tag displays only when user is logged-in -->
-      <h1 *ngIf="authService.user | async">
-        Welcome User: {{ (authService.user | async)?.email }}!
-      </h1>
-
-      <!-- this div displays only when user is not-logged-in -->
-      <div *ngIf="!(authService.user | async)">
-        <!-- these input fields takes email and password from user -->
-        <input type="text" [(ngModel)]="email" placeholder="Enter email"/>
-        <input type="password" [(ngModel)]="password" placeholder="Enter password"/>
-
-        <!-- user click this button to signup -->
-        <button (click)="signup()" [disabled]="!email || !password">
-          Signup
-        </button>
-
-        <!-- user click this button to login -->
-        <button (click)="login()" [disabled]="!email || !password">
-          Login
-        </button>
-      </div>
-
-      <!-- user click this button to logout -->
-      <!-- this button displays only when user is logged-in -->
-      <button (click)="logout()" *ngIf="authService.user | async">
-        Logout
-      </button>
-    </div>
-
-
-
     <exp-loader></exp-loader>
 
     <router-outlet></router-outlet>
