+++ src/app/shared/components/registration/registration.component.ts
@@ -2,8 +2,8 @@ import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 
-import { AuthService } from '../../services/auth.service';
 import { MessagesService } from '../../services/messages.service';
+import { AuthService } from '../../../service/auth.service';
 
 @Component({
   selector: 'exp-registration',
@@ -16,60 +16,78 @@ export class RegistrationComponent implements OnInit {
   public userRoles = ['admin', 'manager', 'HR'];
   public role: string;
 
+  public email: string;
+  public password: string;
+
   private editInProgress = false;
 
   constructor(
-    private authService: AuthService,
+    public authService: AuthService,
     private router: Router,
     private formBuilder: FormBuilder,
     private msgService: MessagesService
   ) { }
 
   public ngOnInit(): void {
-    this.initRegisterForm();
-
-    const isLogin = this.authService.isLoggedIn();
-
-    if (isLogin) {
-      this.router.navigate(['/main']);
-    }
+    // this.initRegisterForm();
+    //
+    // const isLogin = this.authService.isLoggedIn();
+    //
+    // if (isLogin) {
+    //   this.router.navigate(['/main']);
+    // }
   }
 
-  private initRegisterForm(): void {
-    this.registerForm = this.formBuilder.group({
-      username:    ['', Validators.required],
-      login:    ['', Validators.required],
-      password: ['', [Validators.required, Validators.minLength(4)]]
-    });
+  // AUTH
+  public signup(): any {
+    this.authService.signup(this.email, this.password);
+    this.email = this.password = '';
   }
 
-  registration(): void {
-    console.log('REGISTRATION: ', this.registerForm.value.username, this.registerForm.value.login, this.registerForm.value.password);
-
-    this.editInProgress = false;
-    this.authService.login(this.registerForm.value.login, this.registerForm.value.password)
-      .subscribe(
-        () => {
-          this.msgService.setMessage({
-            type: 'success',
-            body: `${this.registerForm.value.login}, You have successfully logged in to your account. Welcome in Film Catalog!`
-          });
-          setTimeout(() => {
-            this.router.navigate(['/main']);
-          }, 2000);
-        },
-        err => {
-          this.msgService.setMessage({
-            type: 'danger',
-            body: err.error.error
-          });
-        }
-      );
+  public login(): any {
+    this.authService.login(this.email, this.password);
+    this.email = this.password = '';
   }
 
-
-  public goToLogin(): void {
-    this.router.navigate(['/login']);
+  public logout(): any {
+    this.authService.logout();
   }
+
+  // private initRegisterForm(): void {
+  //   this.registerForm = this.formBuilder.group({
+  //     username:    ['', Validators.required],
+  //     login:    ['', Validators.required],
+  //     password: ['', [Validators.required, Validators.minLength(4)]]
+  //   });
+  // }
+  //
+  // registration(): void {
+  //   console.log('REGISTRATION: ', this.registerForm.value.username, this.registerForm.value.login, this.registerForm.value.password);
+  //
+  //   this.editInProgress = false;
+  //   this.authService.login(this.registerForm.value.login, this.registerForm.value.password)
+  //     .subscribe(
+  //       () => {
+  //         this.msgService.setMessage({
+  //           type: 'success',
+  //           body: `${this.registerForm.value.login}, You have successfully logged in to your account. Welcome in Film Catalog!`
+  //         });
+  //         setTimeout(() => {
+  //           this.router.navigate(['/main']);
+  //         }, 2000);
+  //       },
+  //       err => {
+  //         this.msgService.setMessage({
+  //           type: 'danger',
+  //           body: err.error.error
+  //         });
+  //       }
+  //     );
+  // }
+  //
+  //
+  // public goToLogin(): void {
+  //   this.router.navigate(['/login']);
+  // }
 
 }
