+++ src/app/shared/guards/auth-guard.service.ts
@@ -2,8 +2,7 @@ import { Injectable } from '@angular/core';
 import { ActivatedRouteSnapshot, CanActivate, Router, RouterStateSnapshot, UrlTree } from '@angular/router';
 
 import { Observable } from 'rxjs';
-
-import { AuthService } from '../services/auth.service';
+import { AuthService } from '../../service/auth.service';
 
 @Injectable({
   providedIn: 'root'
